requirements.txt
ipdb==0.10.2
ipython==5.3.0
Django==1.10.5
djangorestframework==3.5.4
MySQL-python==1.2.5
pytz==2015.7